fix: clean up and connect progress bar
