+++ packages/client/hmi-client/src/components/workflow/tera-calibrate-ensemble-node-ciemss.vue
@@ -53,8 +53,7 @@ import {
 	makeEnsembleCiemssCalibration,
 	getRunResultCiemss,
 	simulationPollAction,
-	querySimulationInProgress,
-	EventSourceManager
+	querySimulationInProgress
 } from '@/services/models/simulation-service';
 import Button from 'primevue/button';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
@@ -94,7 +93,6 @@ const datasetColumnNames = ref<string[]>();
 const progress = ref({ status: ProgressState.RETRIEVING, value: 0 });
 
 const poller = new Poller();
-const eventSourceManager = new EventSourceManager();
 
 onMounted(() => {
 	const runIds = querySimulationInProgress(props.node);
@@ -138,9 +136,7 @@ const getStatus = async (simulationId: string) => {
 	poller
 		.setInterval(3000)
 		.setThreshold(300)
-		.setPollAction(async () =>
-			simulationPollAction(runIds, props.node, progress, emit, eventSourceManager)
-		);
+		.setPollAction(async () => simulationPollAction(runIds, props.node, progress, emit));
 	const pollerResults = await poller.start();
 
 	if (pollerResults.state !== PollerState.Done || !pollerResults.data) {
