+++ packages/client/hmi-client/src/components/workflow/tera-calibration-node-ciemss.vue
@@ -118,16 +118,14 @@ import {
 	makeCalibrateJobCiemss,
 	getRunResultCiemss,
 	simulationPollAction,
-	querySimulationInProgress,
-	EventSourceManager
+	querySimulationInProgress
 } from '@/services/models/simulation-service';
 import { setupModelInput, setupDatasetInput } from '@/services/calibrate-workflow';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
 import { workflowEventBus } from '@/services/workflow';
 import _ from 'lodash';
 import { Poller, PollerState } from '@/api/api';
-import useAuthStore from '@/stores/auth';
-import { EventSourcePolyfill } from 'event-source-polyfill';
+import { EventSourceManager } from '@/utils/event-source-manager';
 import {
 	CalibrationOperationCiemss,
 	CalibrationOperationStateCiemss,
@@ -252,30 +250,6 @@ const runCalibrate = async () => {
 	};
 	const response = await makeCalibrateJobCiemss(calibrationRequest);
 
-	// Run MQ Test
-	console.log(`Checking MQ for job Id ${response.simulationId}`);
-	const auth = useAuthStore();
-	const eventsource = new EventSourcePolyfill(
-		`/api/simulations/${response.simulationId}/partial-result`,
-		{
-			headers: {
-				Authorization: `Bearer ${auth.token}`
-			}
-		}
-	);
-	eventsource.onerror = (e) => {
-		console.log('An error occurred while attempting to connect.');
-		console.log(e);
-		eventsource.close();
-		console.log('Closing eventsource');
-	};
-	eventsource.onmessage = (event) => {
-		console.log(event);
-		const json = event.data;
-		console.log(json);
-		eventsource.close();
-		console.log('Closing eventsource');
-	};
 	if (response?.simulationId) {
 		getStatus(response.simulationId);
 	}
@@ -290,14 +264,29 @@ const getStatus = async (simulationId: string) => {
 	}
 	console.log(`Simulation Id:${simulationId}`);
 	const runIds = [simulationId];
+
+	// open a connection for each run id and handle the messages
+	runIds.forEach((id) => {
+		eventSourceManager.openConnection(id);
+		eventSourceManager.setMessageHandler(id, (message) => {
+			const parsedMessage = JSON.parse(message);
+			if (parsedMessage.progres) {
+				progress.value.value = Math.round(parsedMessage.progress * 100);
+			}
+		});
+	});
+
 	poller
 		.setInterval(3000)
 		.setThreshold(300)
-		.setPollAction(async () =>
-			simulationPollAction(runIds, props.node, progress, emit, eventSourceManager)
-		);
+		.setPollAction(async () => simulationPollAction(runIds, props.node, progress, emit));
 	console.log('Poller defined');
 	const pollerResults = await poller.start();
+
+	// closing event source connections
+	runIds.forEach((id) => {
+		eventSourceManager.closeConnection(id);
+	});
 
 	if (pollerResults.state !== PollerState.Done || !pollerResults.data) {
 		// throw if there are any failed runs for now
