+++ packages/client/hmi-client/src/components/workflow/tera-simulate-julia-node.vue
@@ -29,8 +29,7 @@ import {
 	makeForecastJob,
 	getRunResult,
 	simulationPollAction,
-	querySimulationInProgress,
-	EventSourceManager
+	querySimulationInProgress
 } from '@/services/models/simulation-service';
 import { ProgressState, WorkflowNode } from '@/types/workflow';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
@@ -56,7 +55,6 @@ const modelConfigId = computed<string | undefined>(() => props.node.inputs[0].va
 const progress = ref({ status: ProgressState.RETRIEVING, value: 0 });
 
 const poller = new Poller();
-const eventSourceManager = new EventSourceManager();
 
 onMounted(() => {
 	const runIds = querySimulationInProgress(props.node);
@@ -99,9 +97,7 @@ const getStatus = async (runIds: string[]) => {
 	poller
 		.setInterval(3000)
 		.setThreshold(300)
-		.setPollAction(async () =>
-			simulationPollAction(runIds, props.node, progress, emit, eventSourceManager)
-		);
+		.setPollAction(async () => simulationPollAction(runIds, props.node, progress, emit));
 	const pollerResults = await poller.start();
 
 	if (pollerResults.state !== PollerState.Done || !pollerResults.data) {
